[...] Misc changes to rendering

+++ Xenon/Render/Dummy/DummyBuffer.h
@@ -17,8 +17,8 @@ public:
     DestroyBuffer();
   }
 
-  void CreateBuffer(u32 size, const void *data, eBufferUsage usage, eBufferType type) override;
-  void UpdateBuffer(u32 offset, u32 size, const void *data) override;
+  void CreateBuffer(u64 size, const void *data, eBufferUsage usage, eBufferType type) override;
+  void UpdateBuffer(u64 offset, u64 size, const void *data) override;
   void Bind(u32 binding) override;
   void Unbind() override;
   void DestroyBuffer() override;
